Feature/add sensor

+++ src/main/java/live/ioteatime/apiservice/dto/OrganizationDto.java
@@ -16,5 +16,5 @@ public class OrganizationDto {
     @Schema(description = "조직 이름입니다.")
     private String name;
     @Schema(description = "조직 ADMIN이 설정한 이번 달 목표 전기 요금 값입니다.")
-    private BigInteger electricityBudget;
+    private Long electricityBudget;
 }
